Move `::test_get_order_id_null_before_set()`
