+++ tests/wpunit/api/addresses/class-bitcoin-address-wpunit-Test.php
@@ -52,32 +52,6 @@ class Bitcoin_Address_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCas
 
 		$this->assertNotEquals( $last_modified_time_before, $last_modified_time_after );
 	}
-
-	/**
-	 * @covers ::get_order_id
-	 */
-	public function test_get_order_id_null_before_set(): void {
-
-		$this->markTestIncomplete( 'should it be possible for a Bitcoin_Address object to exist without a post_id?' );
-
-		$bitcoin_address_factory    = new Bitcoin_Address_Factory();
-		$bitcoin_address_repository = new Bitcoin_Address_Repository( $bitcoin_address_factory );
-
-		$wallet = $this->makeEmpty( Bitcoin_Wallet::class );
-
-		$bitcoin_address = $bitcoin_address_repository->save_new_address(
-			wallet:             $wallet,
-			derivation_path_sequence_index: 2,
-			address: 'address',
-		);
-
-		$sut = $bitcoin_address_repository->get_by_post_id( $bitcoin_address->get_post_id() );
-
-		$result = $sut->get_order_id();
-
-		$this->assertNull( $result );
-	}
-
 	/**
 	 * @covers ::get_order_id
 	 */
